giga segg buff
- added response to mi lesz
- added response to segged kivan
- renamed file and class to better represent new features
