+++ cog_handler.py
@@ -1,4 +1,4 @@
-from cogs import ping, ghostping, asked, react, symouse, insult, leseird, nemertem, faszopo, dummy_police, balint_reacting, mivan
+from cogs import ping, ghostping, asked, react, symouse, insult, leseird, nemertem, faszopo, dummy_police, balint_reacting, segg
 
 async def add_all_cogs(bot):
     await bot.add_cog(ping.Ping(bot))
@@ -16,4 +16,4 @@ async def add_all_cogs(bot):
     await bot.add_cog(faszopo.Faszopo(bot))
     await bot.add_cog(dummy_police.DummyPolice(bot))
     await bot.add_cog(balint_reacting.BalintReacting(bot))
-    await bot.add_cog(mivan.Mivan(bot))
+    await bot.add_cog(segg.Segg(bot))
